requirements.txt
torch
spacy
ruamel.yaml
numpy
matplotlib
scipy
networkx
tqdm
